fix:long length

+++ library/asn.1/per_support.h
@@ -25,7 +25,7 @@ typedef struct asn_per_constraint_s {
 	int  range_bits;		/* Full number of bits in the range */
 	int  effective_bits;		/* Effective bits */
 	long lower_bound;		/* "lb" value */
-	long upper_bound;		/* "ub" value */
+	long long upper_bound;		/* "ub" value */
 } asn_per_constraint_t;
 typedef struct asn_per_constraints_s {
 	asn_per_constraint_t value;
